[core] Merge vector_tile_data.{hpp,cpp} into vector_tile.{hpp,cpp}
