+++ src/mbgl/tile/vector_tile.hpp
@@ -1,7 +1,10 @@
 #pragma once
 
 #include <mbgl/tile/geometry_tile.hpp>
+#include <mbgl/tile/geometry_tile_data.hpp>
 #include <mbgl/tile/tile_id.hpp>
+#include <mbgl/tile/tile_source.hpp>
+
 #include <protozero/pbf_reader.hpp>
 
 #include <map>
@@ -9,6 +12,29 @@
 #include <functional>
 
 namespace mbgl {
+
+class Tileset;
+class GeometryTile;
+
+namespace style {
+class UpdateParameters;
+}
+
+class VectorTileData : public GeometryTileData {
+public:
+    VectorTileData(const OverscaledTileID&,
+                   std::string sourceID,
+                   const style::UpdateParameters&,
+                   const Tileset&);
+
+    void setNecessity(Necessity) final;
+    void setData(std::shared_ptr<const std::string> data,
+                 optional<Timestamp> modified,
+                 optional<Timestamp> expires);
+
+private:
+    TileSource<VectorTileData> tileSource;
+};
 
 class VectorTileLayer;
